requirements.txt
# Django & Django-related packages
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
django-tailwind==4.0.1
django-environ==0.12.0

# Database
psycopg2-binary==2.9.9

# Production server
gunicorn==21.2.0
whitenoise==6.7.0

# WSGI server
uvicorn[standard]==0.29.0

# Security
python-dotenv==1.1.0

# MCP related
modelcontextprotocol==0.1.0
